+++ spec/lib/openapi3_parser/node/schema_spec.rb
@@ -57,5 +57,35 @@ RSpec.describe Openapi3Parser::Node::Schema do
         expect(node.requires?(node.properties["field_b"])).to be false
       end
     end
+
+    context "when comparing referenced schemas" do
+      let(:node) do
+        input = {
+          "type" => "object",
+          "required" => %w[field_a],
+          "properties" => {
+            "field_a" => { "$ref" => "#/referenced_item" },
+            "field_b" => { "$ref" => "#/referenced_item" }
+          }
+        }
+
+        document_input = {
+          "referenced_item" => { "type" => "string" }
+        }
+
+        factory_context = create_node_factory_context(input, document_input: document_input)
+        Openapi3Parser::NodeFactory::Schema
+          .new(factory_context)
+          .node(node_factory_context_to_node_context(factory_context))
+      end
+
+      it "returns true for the required reference field" do
+        expect(node.requires?(node.properties["field_a"])).to be true
+      end
+
+      it "returns false for the reference field that isn't required" do
+        expect(node.requires?(node.properties["field_b"])).to be false
+      end
+    end
   end
 end